Move test of optimization into test_opt.py
