+++ theano/gpuarray/tests/test_opt.py
@@ -19,7 +19,7 @@ from ..elemwise import GpuCAReduceCuda, GpuCAReduceCPY, GpuElemwise
 from ..subtensor import GpuSubtensor
 from ..linalg import GpuCusolverSolve, cusolver_available
 
-from .config import mode_with_gpu, test_ctx_name, SkipTest
+from .config import mode_with_gpu, mode_without_gpu, test_ctx_name, SkipTest
 
 
 def test_local_assert():
@@ -446,6 +446,51 @@ def test_local_gpu_elemwise():
     out = f(a_v, b_v)
     utt.assert_allclose(out[0], a_v[::2] + b_v[::2])
     utt.assert_allclose(out[1], a_v[::2] * c_v[::2])
+
+
+def test_many_arg_elemwise():
+    # this test checks whether the + and * elemwise ops can handle
+    # extremely large numbers of arguments on gpu
+    rng = np.random.RandomState([1, 2, 3])
+
+    for num_args in [75]:
+        for op_to_test in [theano.tensor.add, theano.tensor.mul]:
+            for nb_dim in [2, 3, 4, 5, 7]:
+                shapes = [rng.randint(1, 5) for i in range(nb_dim)]
+                args = [np.cast['float32'](rng.randn(*shapes))
+                        for arg in range(0, num_args)]
+
+                symb_args = [theano.tensor.TensorType('float32',
+                                                      (False,) * nb_dim)()
+                             for arg in range(0, num_args)]
+
+                outputs = []
+                for mode in [mode_with_gpu, mode_without_gpu]:
+                    # test the optijmization local_gpu_elemwise_0
+                    f = theano.function(
+                        symb_args, op_to_test(*symb_args),
+                        mode=mode.excluding("local_gpu_elemwise_1"))
+                    outputs.append(f(*args))
+                    # assert that the test was done on the gpu.
+                    if mode is mode_with_gpu:
+                        assert any([isinstance(node.op, GpuElemwise)
+                                    for node in f.maker.fgraph.apply_nodes])
+
+                    # test the optijmization local_gpu_elemwise_1
+                    f = theano.function(
+                        symb_args,
+                        GpuFromHost(test_ctx_name)(op_to_test(*symb_args)),
+                        mode=mode.excluding("local_gpu_elemwise_0"))
+                    out = f(*args)
+                    # assert that the test was done on the gpu.
+                    if mode is mode_with_gpu:
+                        assert any([isinstance(node.op, GpuElemwise)
+                                    for node in f.maker.fgraph.apply_nodes])
+                    utt.assert_allclose(out, outputs[-1])
+
+                results_gpu, results_cpu = outputs
+
+                utt.assert_allclose(results_gpu, results_cpu)
 
 
 def test_local_lift_abstractconv_gpu_shape():
